OSSL_FN: Add OSSL_FN_clear_bit()
The OSSL_FN counterpart of BN_clear_bit(): clear a bit by absolute
index, failing with OSSL_FN_R_RESULT_ARG_TOO_SMALL when the index is
out of range (OSSL_FN is fixed-size, so the operand cannot be grown to
reach it).  Control flow branches only on the operand's public width
and the caller-chosen index, never on limb values, matching the leak
profile of the BN counterpart.

The bit index is size_t rather than the BN counterpart's int; a
negative index was never meaningful (an error for clearing, a
read-as-0 for testing).  OSSL_FN_is_bit_set() changes along, to keep
the bit-index family consistent.

Related-to: doc/designs/fixed-size-large-numbers.md
Issue: https://github.com/openssl/project/issues/1654

+++ crypto/fn/fn_rand.c
@@ -180,8 +180,8 @@ static int ossl_fn_rand_range(enum ossl_fn_rand_flag flag, OSSL_FN *r,
 
     if (n == 1) {
         return OSSL_FN_zero(r);
-    } else if (!OSSL_FN_is_bit_set(range, (int)(n - 2))
-        && !OSSL_FN_is_bit_set(range, (int)(n - 3))
+    } else if (!OSSL_FN_is_bit_set(range, n - 2)
+        && !OSSL_FN_is_bit_set(range, n - 3)
         && n < (size_t)r->dsize * OSSL_FN_BITS) {
         /*
          * range = 100..._2, so 3*range (= 11..._2) is exactly one bit longer
